many fixes and redesigns

+++ lib/ksnutils/src/memset.cpp
@@ -4,7 +4,7 @@ extern "C"
 {
     void *memset(void *bufptr, int value, size_t size)
     {
-        set_memory(reinterpret_cast<uint8_t *>(bufptr), value, size);
+        set_memory(reinterpret_cast<uint8_t *>(bufptr), static_cast<uint8_t>(value), size);
         return bufptr;
     }
 }
